modify format and watched video screen
